tickets/src/routes/update.ts
import express, { Request, Response } from 'express'
import {
  requireAuth,
  validateRequest,
  NotFoundError,
  NotAuthorizedError,
} from '@mveer/common'
import { body } from 'express-validator'
import { Ticket } from '../models/ticket'
const router = express.Router()

router.put(
  '/api/tickets/:id',
  requireAuth,
  [
    body('title').not().isEmpty().withMessage('Title required'),
    body('price').isFloat({ gt: 0 }).withMessage('Price > 0 required'),
  ],
  validateRequest,
  async (req: Request, res: Response) => {
    const ticket = await Ticket.findById(req.params.id)

    if (!ticket) {
      throw new NotFoundError()
    }

    if (ticket.userId !== req.currentUser?.id) {
      throw new NotAuthorizedError()
    }

    res.send(ticket)
  },
)

export { router as updateTicketRouter }
